test: split shared AGY fixture helpers
